EDM-1151: Automation of policy update enforcement
